requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0
python-multipart==0.0.6
huggingface_hub
requests==2.31.0
beautifulsoup4==4.12.3
